tests/wpt/web-platform-tests/content-security-policy/style-src-attr-elem/style-src-elem-blocked-src-allowed.html

<!DOCTYPE html>
<html>

<head>
    <meta http-equiv="Content-Security-Policy" content="style-src-elem 'none';
    style-src 'unsafe-inline';">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script>
      var t = async_test("Should fire a security policy violation event");
      window.addEventListener('securitypolicyviolation', t.step_func_done(function(e) {
        assert_equals(e.violatedDirective, 'style-src-elem');
        assert_equals(e.blockedURI, 'inline');
      }));
    </script>
</head>

<body>
  <style>
    body {background: green;}
  </style>
  <script>
    async_test(function(test) {
      assert_equals(document.styleSheets.length, 0);
      test.done();
    }, "The inline style should not be applied");
  </script>
</body>

</html>
